test(app): move the recorder doubles out of the capped helpers file
`TestHelpers.swift` landed on 606 lines after the rebase, six past the cap
that `--strict` turns into an error. Neither side broke it alone: this
branch adds nine lines to that file and main has since added a second
recorder double to it.

The doubles are what moves, because they are the coherent piece: each one
stands in for the recorder at the same seam, and a test picks between them
by the outcome it needs — a stop that completes, or a start that cannot
open. The file is named after `MockRecorder` so the lint rule that wants a
file named for a type it declares is satisfied.

No behaviour, no test changed; the factory moved with the type it builds.
